fix: make SpawnId null.Value[int64], not [uint64]
Review finding: sql.Null[uint64].Value() delegates to
driver.DefaultParameterConverter.ConvertValue, which explicitly refuses
any value with the high bit set ("uint64 values with high bit set are
not supported"). The nulltypes.NullUint64.Value() this replaced did a
bit-reinterpret (return int64(n.V), nil) that always succeeded, so the
type swap silently dropped a real capability: null.Value[uint64] can
Scan the full unsigned range of spawn_id (bigint unsigned) but cannot
always Value() it back. Unreachable today only because both
SetSpawnId call sites route through strconv.ParseInt(s, 16, 64), which
can't produce a sign-bit value — an upstream accident, not a proof of
equivalence.

Switches SpawnId to null.Value[int64], matching CellId's existing
type and reasoning: int64 has no such restriction, and real spawn ids
are ~48-bit, far below where either type's range would matter.
SetSpawnId now stores v.Int64 directly with no cast, same shape as
SetCellId. Downstream: pokemon_state.go's
strconv.FormatUint(..., 16) call picked up an explicit uint64(...)
cast (matching the pattern already used for CellId at pokemon_state.go
:282), and three call sites building *int64 API fields
(api_pokemon_response.go, api_pokemon_scan_v2.go, api_pokemon_scan_v3.go)
simplified from widenPtr[uint64, int64](...) to a plain .Ptr(), since
SpawnId.Ptr() now already returns *int64.

Also corrects a gap the previous commit's report should have caught:
TestPokemonFullRowRoundTrip never actually set or asserted SpawnId
against the live database - the value it cited as proof
(SpawnId: null.ValueFrom(uint64(7777))) lives in the unrelated
JSON-only golden test in api_pokemon_response_test.go, which never
touches the database. Added a real SpawnId round-trip assertion here
so the type this commit picks is actually exercised against MariaDB,
not just proven to compile.

Sizes unchanged (PokemonData=256, Pokemon=392): int64 and uint64 are
both 8-byte payloads, so null.Value[int64] and null.Value[uint64] are
both 16 bytes.

+++ decoder/pokemon_decode.go
@@ -489,7 +489,7 @@ func (pokemon *Pokemon) setExpireTimestampFromSpawnpoint(ctx context.Context, db
 		return
 	}
 
-	spawnId := int64(pokemon.SpawnId.ValueOrZero())
+	spawnId := pokemon.SpawnId.ValueOrZero()
 	if spawnId == 0 {
 		return
 	}
